+++ src/sst/core/Makefile.am
@@ -66,6 +66,7 @@ nobase_dist_sst_HEADERS = \
 	sparseVectorMap.h \
 	sst_types.h \
 	sstinfo.h \
+	sstpart.h \
 	stopAction.h \
 	stringize.h \
 	subcomponent.h \
