+++ src/sst/core/impl/Makefile.inc
@@ -4,3 +4,4 @@
 #
 
 include impl/timevortex/Makefile.inc
+include impl/partitioners/Makefile.inc
