+++ src/sst/core/impl/partitioners/rrobin.h
@@ -1,61 +1,52 @@
 // Copyright 2009-2017 Sandia Corporation. Under the terms
 // of Contract DE-NA0003525 with Sandia Corporation, the U.S.
 // Government retains certain rights in this software.
-//
+// 
 // Copyright (c) 2009-2017, Sandia Corporation
 // All rights reserved.
-//
+// 
 // This file is part of the SST software package. For license
 // information, see the LICENSE file in the top level directory of the
 // distribution.
+#ifndef SST_CORE_IMPL_PARTITONERS_RROBIN_H
+#define SST_CORE_IMPL_PARTITONERS_RROBIN_H
 
-
-#ifndef SST_CORE_PART_SINGLE
-#define SST_CORE_PART_SINGLE
-
-#include <sst/core/part/sstpart.h>
+#include <sst/core/sstpart.h>
 #include <sst/core/elementinfo.h>
 
-using namespace SST;
-using namespace SST::Partition;
-
 namespace SST {
+namespace IMPL {
 namespace Partition {
 
-
-    
-/**
-   Single partitioner is a virtual partitioner used for serial jobs.
-   It simply ensures that all components are assigned to rank 0.
-*/
-class SSTSinglePartition : public SST::Partition::SSTPartitioner {
+class SSTRoundRobinPartition : public SST::Partition::SSTPartitioner {
 
 public:
     SST_ELI_REGISTER_PARTITIONER(
-        SSTSinglePartition,
+        SSTRoundRobinPartition,
         "sst",
-        "single",
+        "roundrobin",
         SST_ELI_ELEMENT_VERSION(1,0,0),
-        "Allocates all components to rank 0.  Automatically selected for serial jobs.")
+        "Partitions components using a simple round robin scheme based on ComponentID.  Sequential IDs will be placed on different ranks.")
 
-    /**
-       Creates a new single partition scheme.
-    */
-    SSTSinglePartition(RankInfo total_ranks, RankInfo my_rank, int verbosity);
+private:
+    RankInfo world_size;
+
+public:
+    SSTRoundRobinPartition(RankInfo world_size, RankInfo my_rank, int verbosity);
     
     /**
        Performs a partition of an SST simulation configuration
        \param graph The simulation configuration to partition
     */
-    void performPartition(PartitionGraph* graph) override;
-    
+	void performPartition(PartitionGraph* graph) override;
+
     bool requiresConfigGraph() override { return false; }
     bool spawnOnAllRanks() override { return false; }
     
-    
+        
 };
 
-}
-}
-
-#endif
+} // namespace Partition
+} // namespace IMPL
+} //namespace SST
+#endif //SST_CORE_IMPL_PARTITONERS_RROBIN_H
