+++ src/sst/core/impl/partitioners/selfpart.cc
@@ -10,4 +10,4 @@
 // distribution.
 
 #include <sst_config.h>
-#include <sst/core/part/selfpart.h>
+#include <sst/core/impl/partitioners/selfpart.h>
