+++ src/sst/core/impl/partitioners/simplepart.h
@@ -8,17 +8,20 @@
 // This file is part of the SST software package. For license
 // information, see the LICENSE file in the top level directory of the
 // distribution.
-#ifndef SST_CORE_PART_SIMPLEPART_H
-#define SST_CORE_PART_SIMPLEPART_H
+#ifndef SST_CORE_IMPL_PARTITONERS_SIMPLEPART_H
+#define SST_CORE_IMPL_PARTITONERS_SIMPLEPART_H
 
 #include <map>
+
 #include <sst/core/sst_types.h>
-#include <sst/core/part/sstpart.h>
+#include <sst/core/sstpart.h>
+
 #include <sst/core/elementinfo.h>
 #include <sst/core/configGraph.h>
 
 namespace SST {
-namespace Partition{
+namespace IMPL {
+namespace Partition {
 
 class SimplePartitioner : public SST::Partition::SSTPartitioner {
 
@@ -58,5 +61,6 @@ public:
 };
 
 } // namespace partition
+} // namespace IMPL
 } //namespace SST
-#endif //SST_CORE_PART_SIMPLERPART_H
+#endif //SST_CORE_IMPL_PARTITONERS_SIMPLERPART_H
