+++ src/sst/core/impl/partitioners/singlepart.h
@@ -1,50 +1,60 @@
 // Copyright 2009-2017 Sandia Corporation. Under the terms
 // of Contract DE-NA0003525 with Sandia Corporation, the U.S.
 // Government retains certain rights in this software.
-// 
+//
 // Copyright (c) 2009-2017, Sandia Corporation
 // All rights reserved.
-// 
+//
 // This file is part of the SST software package. For license
 // information, see the LICENSE file in the top level directory of the
 // distribution.
-#ifndef SST_CORE_PART_RROBIN_H
-#define SST_CORE_PART_RROBIN_H
 
-#include <sst/core/part/sstpart.h>
+
+#ifndef SST_CORE_IMPL_PARTITONERS_SINGLEPART_H
+#define SST_CORE_IMPL_PARTITONERS_SINGLEPART_H
+
+#include <sst/core/sstpart.h>
 #include <sst/core/elementinfo.h>
 
 namespace SST {
+namespace IMPL {
 namespace Partition {
 
-class SSTRoundRobinPartition : public SST::Partition::SSTPartitioner {
+
+    
+/**
+   Single partitioner is a virtual partitioner used for serial jobs.
+   It simply ensures that all components are assigned to rank 0.
+*/
+class SSTSinglePartition : public SST::Partition::SSTPartitioner {
 
 public:
     SST_ELI_REGISTER_PARTITIONER(
-        SSTRoundRobinPartition,
+        SSTSinglePartition,
         "sst",
-        "roundrobin",
+        "single",
         SST_ELI_ELEMENT_VERSION(1,0,0),
-        "Partitions components using a simple round robin scheme based on ComponentID.  Sequential IDs will be placed on different ranks.")
+        "Allocates all components to rank 0.  Automatically selected for serial jobs.")
 
-private:
-    RankInfo world_size;
-
-public:
-    SSTRoundRobinPartition(RankInfo world_size, RankInfo my_rank, int verbosity);
+    /**
+       Creates a new single partition scheme.
+    */
+    SSTSinglePartition(RankInfo total_ranks, RankInfo my_rank, int verbosity);
     
     /**
        Performs a partition of an SST simulation configuration
        \param graph The simulation configuration to partition
     */
-	void performPartition(PartitionGraph* graph) override;
-
+    void performPartition(PartitionGraph* graph) override;
+    
     bool requiresConfigGraph() override { return false; }
     bool spawnOnAllRanks() override { return false; }
     
-        
+    
 };
 
-} // namespace Partition
-} //namespace SST
-#endif //SST_CORE_PART_RROBIN_H
+}
+}
+}
+
+#endif
